Fixes animation bug

+++ src/game/Game.cpp
@@ -47,7 +47,7 @@ namespace jumper
 
     void Game::spawnBots()
     {
-        int curPos = m_level->m_camera.x() + m_level->m_camera.w();
+        int curPos = ((int) m_level->m_camera.x()) + m_level->m_camera.w();
 
         vector<Bot*> erease_bots;
 
@@ -89,9 +89,7 @@ namespace jumper
 
     Level* Game::getLevel()
     {
-
         return m_level;
-
     }
 
     void Game::addActor(Actor* actor)
